requirements.txt
DateTime==5.2
Flask==2.3.3
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
line-bot-sdk==3.5.0
Pillow==10.1.0
PyGithub==2.1.1
pyimgur==0.6.0
requests==2.31.0
selenium==4.14.0
urllib3==2.0.5
webdriver-manager
